Fix: fix the compilation failure with new DFTD3 parameter autoset
