+++ source/module_hamilt_general/module_vdw/vdwd3_parameters.cpp
@@ -7,7 +7,6 @@
 #include "vdwd3_parameters.h"
 #include "module_base/constants.h"
 #include <map>
-#include "dftd3_xc_param.h"
 namespace vdw
 {
 
@@ -24,10 +23,10 @@ void Vdwd3Parameters::initial_parameters(const Input_para &input, std::ofstream*
                          5,
                          std::vector<std::vector<double>>(max_elem_, std::vector<double>(max_elem_, 0.0)))));
     
-    DFTD3::dftd3_params(input.dft_functional, input.vdw_method,
-                        input.vdw_s6, input.vdw_s8, input.vdw_a1, input.vdw_a2,
-                        s6_, s18_, rs6_, rs18_, /* rs6: a1, rs18: a2 */
-                        plog);
+    dftd3_params(input.dft_functional, input.vdw_method,
+                 input.vdw_s6, input.vdw_s8, input.vdw_a1, input.vdw_a2,
+                 s6_, s18_, rs6_, rs18_, /* rs6: a1, rs18: a2 */
+                 plog);
     abc_ = input.vdw_abc;
     version_ = input.vdw_method;
     model_ = input.vdw_cutoff_type;
